backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pillow==10.2.0
pytesseract==0.3.10
pdfplumber==0.10.3
PyPDF2==3.0.1
reportlab==4.0.9
openai==1.54.0
httpx==0.27.2
httpcore==1.0.6
h11==0.14.0
python-dotenv==1.0.0
pdf2image==1.17.0
numpy==1.24.3
pandas==2.0.3
